third_party/WebKit/LayoutTests/editing/style/remove-underline-from-stylesheet.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    // Test for http://wkb.ug/27809
    test(() => assert_selection(
        [
            '<style>.underline { text-decoration: underline; }</style>',
            '<div class="underline" contenteditable>^abc|</div>',
        ].join(''),
        'underline',
        [
            '<style>.underline { text-decoration: underline; }</style>',
            '<div class="underline" contenteditable>^abc|</div>',
        ].join('')),
        'Underline command does nothing when selected text has undeline (style sheet).');

    test(() => assert_selection(
        '<div contenteditable style="text-decoration: underline">^abc|</div>',
        'underline',
        '<div contenteditable style="text-decoration: underline">^abc|</div>'),
        'Underline command does nothing when selected text has undeline (inline).');
</script>
